Provided a way to retrieve the total usage as a numeric value in GB.

+++ trunk/src/main/java/com/googlecode/pondskum/gui/swing/notifyer/DefaultDisplayDetailsPack.java
@@ -23,8 +23,8 @@ public final class DefaultDisplayDetailsPack extends AbstractDisplayDetailsPack 
     }
 
     @Override
-    public Double getQuotaLimit() {
-        return 100d;
+    public Integer getQuotaLimit() {
+        return 100;
     }
 
     @Override
